+++ libpsx-rs/src/backends/video/opengl/shaders.rs
@@ -6,18 +6,16 @@ use std::ffi::CString;
 
 pub(crate) mod vertex {
     pub(crate) const CRTC: &'static str = include_str!("./shaders/crtc.vert");
-    pub(crate) const SOLID_LINE_LOOP: &'static str = include_str!("./shaders/solid_line_loop.vert");
-    pub(crate) const SOLID_POLYGON: &'static str = include_str!("./shaders/solid_polygon.vert");
-    pub(crate) const SHADED_POLYGON: &'static str = include_str!("./shaders/shaded_polygon.vert");
-    pub(crate) const TEXTURED_POLYGON: &'static str = include_str!("./shaders/textured_polygon.vert");
+    pub(crate) const LINE_LOOP: &'static str = include_str!("./shaders/LINE_LOOP.vert");
+    pub(crate) const SHADED_TRIANGLES: &'static str = include_str!("./shaders/shaded_triangles.vert");
+    pub(crate) const TEXTURED_TRIANGLES: &'static str = include_str!("./shaders/textured_triangles.vert");
 }
 
 pub(crate) mod fragment {
     pub(crate) const CRTC: &'static str = include_str!("./shaders/crtc.frag");
-    pub(crate) const SOLID_LINE_LOOP: &'static str = include_str!("./shaders/solid_line_loop.frag");
-    pub(crate) const SOLID_POLYGON: &'static str = include_str!("./shaders/solid_polygon.frag");
-    pub(crate) const SHADED_POLYGON: &'static str = include_str!("./shaders/shaded_polygon.frag");
-    pub(crate) const TEXTURED_POLYGON: &'static str = include_str!("./shaders/textured_polygon.frag");
+    pub(crate) const LINE_LOOP: &'static str = include_str!("./shaders/LINE_LOOP.frag");
+    pub(crate) const SHADED_TRIANGLES: &'static str = include_str!("./shaders/shaded_triangles.frag");
+    pub(crate) const TEXTURED_TRIANGLES: &'static str = include_str!("./shaders/textured_triangles.frag");
 }
 
 lazy_static! {
