+++ libpsx-rs/src/backends/video/opengl/shaders/line_loop.vert
