--- libpsx-rs/src/backends/video/opengl/shaders/shaded_polygon.frag
@@ -1,8 +1,0 @@
-#version 450 core
-
-layout(location = 0) in vec4 in_color;
-layout(location = 0) out vec4 out_color;
-
-void main() {
-    out_color = in_color;
-}
