+++ libpsx-rs/src/backends/video/opengl/shaders/shaded_triangles.vert
