--- libpsx-rs/src/backends/video/opengl/shaders/solid_polygon.vert
@@ -1,7 +1,0 @@
-#version 450 core
-
-layout(location = 0) in vec2 in_coord;
-
-void main() {
-    gl_Position = vec4(in_coord, 0.0, 1.0);
-}
