+++ libpsx-rs/src/backends/video/opengl/shaders/textured_triangles.vert
