+++ libpsx-rs/src/system/gpu/controllers/backend_dispatch.rs
@@ -7,8 +7,10 @@ use crate::{
     backends::video::VideoBackend,
     system::{
         gpu::types::{
-            rendering::ClutKind,
-            TransparencyMode,
+            rendering::{
+                ClutKind,
+                TransparencyKind,
+            },
         },
         types::ControllerResult,
     },
@@ -18,76 +20,16 @@ use crate::{
     },
 };
 
-pub(crate) fn draw_polygon_3_solid(video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 3], color: Color) -> ControllerResult<Result<(), ()>> {
+pub(crate) fn draw_triangles_shaded(video_backend: &VideoBackend, indices: &[u32], positions: &[Point2D<f32, Normalized>], colors: &[Color], transparency_kind: TransparencyKind) -> ControllerResult<Result<(), ()>> {
     match video_backend {
         VideoBackend::None => Ok(Err(())),
         #[cfg(opengl)]
-        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_polygon_3_solid(backend_params, positions, color)?)),
+        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_triangles_shaded(backend_params, indices, positions, colors, transparency_kind)?)),
         _ => unimplemented!(),
     }
 }
 
-pub(crate) fn draw_polygon_3_transparent(
-    video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 3], color: Color, transparency: TransparencyMode,
-) -> ControllerResult<Result<(), ()>> {
-    match video_backend {
-        VideoBackend::None => Ok(Err(())),
-        #[cfg(opengl)]
-        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_polygon_3_transparent(backend_params, positions, color, transparency)?)),
-        _ => unimplemented!(),
-    }
-}
-
-pub(crate) fn draw_polygon_4_solid(video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 4], color: Color) -> ControllerResult<Result<(), ()>> {
-    match video_backend {
-        VideoBackend::None => Ok(Err(())),
-        #[cfg(opengl)]
-        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_polygon_4_solid(backend_params, positions, color)?)),
-        _ => unimplemented!(),
-    }
-}
-
-pub(crate) fn draw_polygon_4_transparent(
-    video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 4], color: Color, transparency: TransparencyMode,
-) -> ControllerResult<Result<(), ()>> {
-    match video_backend {
-        VideoBackend::None => Ok(Err(())),
-        #[cfg(opengl)]
-        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_polygon_4_transparent(backend_params, positions, color, transparency)?)),
-        _ => unimplemented!(),
-    }
-}
-
-pub(crate) fn draw_polygon_4_textured_framebuffer(
-    video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 4], texcoords: [Point2D<f32, TexcoordNormalized>; 4], clut_kind: ClutKind,
-) -> ControllerResult<Result<(), ()>> {
-    match video_backend {
-        VideoBackend::None => Ok(Err(())),
-        #[cfg(opengl)]
-        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_polygon_4_textured_framebuffer(backend_params, positions, texcoords, clut_kind)?)),
-        _ => unimplemented!(),
-    }
-}
-
-pub(crate) fn draw_polygon_3_shaded(video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 3], colors: [Color; 3]) -> ControllerResult<Result<(), ()>> {
-    match video_backend {
-        VideoBackend::None => Ok(Err(())),
-        #[cfg(opengl)]
-        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_polygon_3_shaded(backend_params, positions, colors)?)),
-        _ => unimplemented!(),
-    }
-}
-
-pub(crate) fn draw_polygon_4_shaded(video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 4], colors: [Color; 4]) -> ControllerResult<Result<(), ()>> {
-    match video_backend {
-        VideoBackend::None => Ok(Err(())),
-        #[cfg(opengl)]
-        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_polygon_4_shaded(backend_params, positions, colors)?)),
-        _ => unimplemented!(),
-    }
-}
-
-pub(crate) fn draw_polygon_4_textured(
+pub(crate) fn draw_triangles_4_textured(
     video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 4], texcoords: [Point2D<f32, TexcoordNormalized>; 4], texture_width: usize, texture_height: usize,
     texture_colors: &[Color],
 ) -> ControllerResult<Result<(), ()>>
@@ -95,7 +37,18 @@ pub(crate) fn draw_polygon_4_textured(
     match video_backend {
         VideoBackend::None => Ok(Err(())),
         #[cfg(opengl)]
-        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_polygon_4_textured(backend_params, positions, texcoords, texture_width, texture_height, texture_colors)?)),
+        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_triangles_4_textured(backend_params, positions, texcoords, texture_width, texture_height, texture_colors)?)),
+        _ => unimplemented!(),
+    }
+}
+
+pub(crate) fn draw_triangles_4_textured_framebuffer(
+    video_backend: &VideoBackend, positions: [Point2D<f32, Normalized>; 4], texcoords: [Point2D<f32, TexcoordNormalized>; 4], clut_kind: ClutKind,
+) -> ControllerResult<Result<(), ()>> {
+    match video_backend {
+        VideoBackend::None => Ok(Err(())),
+        #[cfg(opengl)]
+        VideoBackend::Opengl(ref backend_params) => Ok(Ok(opengl::draw_triangles_4_textured_framebuffer(backend_params, positions, texcoords, clut_kind)?)),
         _ => unimplemented!(),
     }
 }
