+++ libpsx-rs/src/system/gpu/controllers/backend_dispatch/opengl/framebuffer.rs
@@ -1,0 +1,135 @@
+use crate::{
+    backends::video::opengl::{
+        rendering::*,
+        *,
+    },
+    system::{
+        gpu::{
+            controllers::backend_dispatch::opengl::{
+                debug,
+            },
+        },
+        types::ControllerResult,
+    },
+    types::{
+        geometry::*,
+    },
+    utilities::array::extract_rectangle,
+};
+use opengl_sys::*;
+
+pub(crate) fn read_framebuffer_5551(backend_params: &BackendParams, origin: Point2D<f32, Normalized>, size: Size2D<f32, Normalized>) -> ControllerResult<Vec<u16>> {
+    use crate::system::gpu::constants::{
+        VRAM_HEIGHT_LINES,
+        VRAM_WIDTH_16B,
+    };
+
+    // TODO: resources are not free'd.
+    static mut FBO_CONTEXT: Option<(GLuint, GLuint)> = None;
+    static mut PROGRAM_CONTEXT: Option<ProgramContext> = None;
+
+    debug::trace_call(stdext::function_name!());
+
+    let downsample_texture_width = VRAM_WIDTH_16B as GLint;
+    let downsample_texture_height = VRAM_HEIGHT_LINES as GLint;
+
+    // Setup a temporary FBO with 1x internal resolution - we need this to extract the exact pixels out of.
+    {
+        let (_context_guard, _context) = backend_params.context.guard();
+
+        unsafe {
+            glFinish();
+
+            if FBO_CONTEXT.is_none() {
+                let mut downsample_fbo = 0;
+                glGenFramebuffers(1, &mut downsample_fbo);
+                glBindFramebuffer(GL_DRAW_FRAMEBUFFER, downsample_fbo);
+
+                let mut downsample_texture = 0;
+                glGenTextures(1, &mut downsample_texture);
+                glBindTexture(GL_TEXTURE_2D, downsample_texture);
+                glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB5_A1 as GLint, downsample_texture_width, downsample_texture_height, 0, GL_RGBA, GL_FLOAT, std::ptr::null());
+                glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE as GLint);
+                glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE as GLint);
+                glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST as GLint);
+                glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST as GLint);
+
+                glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, downsample_texture, 0);
+                assert!(glCheckFramebufferStatus(GL_DRAW_FRAMEBUFFER) == GL_FRAMEBUFFER_COMPLETE);
+
+                FBO_CONTEXT = Some((downsample_fbo, downsample_texture));
+            }
+
+            if PROGRAM_CONTEXT.is_none() {
+                let positions_flat: [f32; 8] = [-1.0, -1.0, 1.0, -1.0, 1.0, 1.0, -1.0, 1.0];
+
+                let texcoords_flat: [f32; 8] = [0.0, 0.0, 1.0, 0.0, 1.0, 1.0, 0.0, 1.0];
+
+                let vs = shaders::compile_shader(shaders::vertex::CRTC, GL_VERTEX_SHADER);
+                let fs = shaders::compile_shader(shaders::fragment::CRTC, GL_FRAGMENT_SHADER);
+                let program = shaders::create_program(&[vs, fs]);
+
+                let mut vao = 0;
+                glGenVertexArrays(1, &mut vao);
+                glBindVertexArray(vao);
+                glEnableVertexAttribArray(0);
+                glEnableVertexAttribArray(1);
+
+                let mut vbo_position = 0;
+                glGenBuffers(1, &mut vbo_position);
+                glBindBuffer(GL_ARRAY_BUFFER, vbo_position);
+                glBufferData(GL_ARRAY_BUFFER, 8 * std::mem::size_of::<f32>() as GLsizeiptr, positions_flat.as_ptr() as *const std::ffi::c_void, GL_STATIC_DRAW);
+                glVertexAttribPointer(0, 2, GL_FLOAT, GL_FALSE as GLboolean, 0, std::ptr::null());
+
+                let mut vbo_texcoord = 0;
+                glGenBuffers(1, &mut vbo_texcoord);
+                glBindBuffer(GL_ARRAY_BUFFER, vbo_texcoord);
+                glBufferData(GL_ARRAY_BUFFER, 8 * std::mem::size_of::<f32>() as GLsizeiptr, texcoords_flat.as_ptr() as *const std::ffi::c_void, GL_STATIC_DRAW);
+                glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE as GLboolean, 0, std::ptr::null());
+
+                PROGRAM_CONTEXT = Some(ProgramContext::new(program, vao, &[vbo_position, vbo_texcoord], &[]));
+            }
+
+            let fbo_context = FBO_CONTEXT.unwrap();
+            glBindFramebuffer(GL_DRAW_FRAMEBUFFER, fbo_context.0);
+            glViewport(0, 0, downsample_texture_width, downsample_texture_height);
+
+            let program_context = PROGRAM_CONTEXT.as_ref().unwrap();
+            glUseProgram(program_context.program_id);
+            glBindVertexArray(program_context.vao_id);
+
+            glBindTexture(GL_TEXTURE_2D, SCENE_TEXTURE);
+            let tex2d_cstr = b"tex2d\0";
+            let tex2d_uniform = glGetUniformLocation(program_context.program_id, tex2d_cstr.as_ptr() as *const GLchar);
+            glUniform1i(tex2d_uniform, 0);
+
+            glDrawArrays(GL_TRIANGLE_FAN, 0, 4);
+            glFinish();
+
+            glBindFramebuffer(GL_DRAW_FRAMEBUFFER, SCENE_FBO);
+            glViewport(0, 0, SCENE_TEXTURE_WIDTH, SCENE_TEXTURE_HEIGHT);
+        }
+    }
+
+    let mut buffer: Vec<u16> = vec![0; VRAM_WIDTH_16B * VRAM_HEIGHT_LINES];
+
+    {
+        let (_context_guard, _context) = backend_params.context.guard();
+
+        unsafe {
+            let fbo_context = FBO_CONTEXT.unwrap();
+            glBindTexture(GL_TEXTURE_2D, fbo_context.1);
+            glGetTexImage(GL_TEXTURE_2D, 0, GL_RGBA, GL_UNSIGNED_SHORT_1_5_5_5_REV, buffer.as_mut_ptr() as *mut std::ffi::c_void);
+        }
+    }
+
+    let pixel_origin_x = ((origin.x + 1.0) * ((VRAM_WIDTH_16B / 2) as f32)) as usize;
+    let pixel_origin_y = ((origin.y + 1.0) * ((VRAM_HEIGHT_LINES / 2) as f32)) as usize;
+    let pixel_origin = Point2D::new(pixel_origin_x, pixel_origin_y);
+
+    let pixel_size_width = ((VRAM_WIDTH_16B / 2) as f32 * size.width) as usize;
+    let pixel_size_height = ((VRAM_HEIGHT_LINES / 2) as f32 * size.height) as usize;
+    let pixel_size = Size2D::new(pixel_size_width, pixel_size_height);
+
+    Ok(extract_rectangle(&buffer, VRAM_WIDTH_16B, pixel_origin, pixel_size))
+}
