+++ libpsx-rs/src/types/array.rs
@@ -1,0 +1,7 @@
+use smallvec::SmallVec;
+
+pub(crate) trait AsFlattened {
+    type Output;
+
+    fn as_flattened(&self) -> SmallVec<[Self::Output; 16]>;
+}
